% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse}
\alias{analyse}
\title{Analyse the Parse Tree of an R Script}
\usage{
analyse(x, path = "")
}
\arguments{
\item{x}{parse tree as returned by \code{\link{parse}}}

\item{path}{for internal use only (when this function is called recursively)}
}
\value{
list representing type information on the nodes in the parse tree
}
\description{
Analyse the Parse Tree of an R Script
}
\examples{
# Parse an R script file (here, a file from kwb.utils)
x <- parse("https://raw.githubusercontent.com/KWB-R/kwb.utils/master/R/log.R")

# Analyse the parse tree (This may take some time!)
result <- kwb.code::analyse(x)

# Show the structure of the result list (only 3 levels!)
str(result, 3)
}
