% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{arg_names}
\alias{arg_names}
\title{Get Argument Names of a Function}
\usage{
arg_names(x)
}
\arguments{
\item{x}{function name or function}
}
\value{
vector of character
}
\description{
Get Argument Names of a Function
}
\examples{
arg_names("sum")
arg_names(mean)
}
