% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_parse_tree.R
\name{extract_from_parse_tree}
\alias{extract_from_parse_tree}
\title{Extract Elements from Parse Tree of R Script}
\usage{
extract_from_parse_tree(
  x,
  matches = matches_function,
  dbg = FALSE,
  path = integer(),
  parent = NULL,
  index = -1
)
}
\arguments{
\item{x}{parse tree as returned by \code{parse}}

\item{matches}{function that is called for each node of the tree. Give a 
function here that returns \code{TRUE} if the object is to be selected and
\code{FALSE} else. The value of \code{TRUE} must be given an attribute 
\code{name} that is expected to be a character of length one.}

\item{dbg}{if \code{TRUE} each node in printed during the climing of the tree}

\item{path}{for internal use}

\item{parent}{for internal use}

\item{index}{for internal use}
}
\value{
vector of character or \code{NULL}
}
\description{
The idea of this function is to collect objects of interest from the parse 
tree, e.g. the names of functions that are called by a script. Therefore,
set the function \code{matches} so that it returns \code{TRUE} for the nodes
in the tree that are of interest.
}
