% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_string_constants.R
\name{find_string_constants}
\alias{find_string_constants}
\title{Show String Constants Used in R Scripts}
\usage{
find_string_constants(root = "./R")
}
\description{
Show String Constants Used in R Scripts
}
