% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elements_by_type.R
\name{get_elements_by_type}
\alias{get_elements_by_type}
\title{Extract Sections of Same "Type" from Parse Tree}
\usage{
get_elements_by_type(x, result = NULL, dbg = TRUE)
}
\arguments{
\item{x}{parse tree as returned by \code{\link{parse}}}

\item{result}{optional. Result as returned by \code{\link{analyse}}}

\item{dbg}{if \code{TRUE}, debug messages are shown}
}
\description{
Extract Sections of Same "Type" from Parse Tree
}
\examples{
# Parse an R script file (here, a file from kwb.utils)
x <- parse("https://raw.githubusercontent.com/KWB-R/kwb.utils/master/R/log.R")

# For each "type" of code segment, extract all occurrences
elements <- get_elements_by_type(x)

# Show all for-loops
elements$`language|call|for|4|`

# Show all if-statements
elements$`language|call|if|3|`

# Show all if-else-statements
elements$`language|call|if|4|`

}
