% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_full_function_info.R
\name{get_full_function_info}
\alias{get_full_function_info}
\title{Get information on function definitions in parsed R scripts}
\usage{
get_full_function_info(trees)
}
\arguments{
\item{trees}{list of R script parse trees as provided by
\code{\link{parse_scripts}}}
}
\description{
Get information on function definitions in parsed R scripts
}
\seealso{
\code{\link{parse_scripts}}
}
