% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_function_assignments.R
\name{get_function_assignments}
\alias{get_function_assignments}
\title{Extract the function assignments from an R script}
\usage{
get_function_assignments(file)
}
\arguments{
\item{file}{path to R script from which function definitions are to be 
extracted}
}
\value{
named list of expressions. The names of the list elements represent 
  the names of the functions that are defined by the expressions in the list.
}
\description{
Extract the function assignments from an R script
}
