% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package_usage_per_script.R
\name{get_package_usage_per_script}
\alias{get_package_usage_per_script}
\title{Get Package Usage per Script}
\usage{
get_package_usage_per_script(root, packages, pattern = "\\\\.R$")
}
\arguments{
\item{root}{root directory with R scripts}

\item{packages}{vector with package names to be checked}

\item{pattern}{default: "\\.R$"}
}
\value{
tibble with information on used packages
}
\description{
Get Package Usage per Script
}
