% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_string_constants_in_scripts.R
\name{get_string_constants_in_scripts}
\alias{get_string_constants_in_scripts}
\title{Get Frequency of String Constant Usage in R Scripts}
\usage{
get_string_constants_in_scripts(root, scripts = dir(root, "\\\\.[Rr]$",
  recursive = TRUE), two_version_check = TRUE)
}
\arguments{
\item{root}{path to folder in which to look for R scripts}

\item{scripts}{optional. Paths to R scripts in which to search for string
constants, relative to \code{root}}

\item{two_version_check}{if \code{TRUE} (default), two different 
implementations of this function are used and the results are compared
internally. Set this argument to \code{FALSE} to get the result as fast
as possible.}
}
\value{
data frame with columns \code{file_id} (file identifier),
  \code{string} (string constant found in the file) and \code{count} (number
  of occurrences of the string counted in the file). The file identifier can
  be resolved to a full file name using the "file database" that is stored in
  the attribute "file_db".
}
\description{
Get Frequency of String Constant Usage in R Scripts
}
\examples{
root <- system.file(package = "kwb.code")
constants <- get_string_constants_in_scripts(root)

# Get paths to files from "file database" stored in attribute "file_db"
kwb.utils::getAttribute(constants, "file_db")

}
