% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{to_full_script_info}
\alias{to_full_script_info}
\title{Get script statistics from a list of R script trees}
\usage{
to_full_script_info(trees)
}
\arguments{
\item{trees}{list of R script parse trees as provided by
\code{\link{parse_scripts}}}
}
\description{
Get script statistics from a list of R script trees
}
\seealso{
\code{\link{parse_scripts}}
}
