% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_weaknesses_in_scripts.R
\name{find_weaknesses_in_scripts}
\alias{find_weaknesses_in_scripts}
\title{Find weaknesses in R scripts}
\usage{
find_weaknesses_in_scripts(x = parse_scripts(root), root = NULL)
}
\arguments{
\item{x}{list of named parse trees as returned by
\code{\link{parse_scripts}}. Not required if \code{root} is given.}

\item{root}{path to folder containing R scripts}
}
\value{
data frame with columns \code{file}, \code{expression},
  \code{frequency}, \code{recommendation}
}
\description{
Find weaknesses in R scripts
}
